Update policy, delete seeder

+++ app/Http/Controllers/SupplierController.php
@@ -13,12 +13,6 @@ class SupplierController extends Controller
     /**
      * Display a listing of the resource.
      */
-
-     public function __construct()
-     {
-        $this->authorizeResource(Supplier::class, 'supplier');
-     }
-
     public function index(Request $request)
     {
         $data = DB::table('suppliers')
@@ -63,6 +57,7 @@ class SupplierController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Supplier::class);
        $validated = $request->validate([
             'name' => 'required',
             'phone' => 'required',
@@ -94,6 +89,7 @@ class SupplierController extends Controller
      */
     public function update(Request $request, Supplier $supplier)
     {
+        $this->authorize('update', Supplier::class);
         $validated = $request->validate([
             'name' => 'required',
             'phone' => 'required',
@@ -108,6 +104,7 @@ class SupplierController extends Controller
      */
     public function destroy(Supplier $supplier)
     {
+        $this->authorize('delete', Supplier::class);
         Supplier::destroy($supplier->id);
     }
 }
